pyproject.toml
[project]
name = "ppt-creator"
version = "0.1.0"
description = "Analyze, Suggest, and Create PPTs automatically."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=2.21.0",
    "python-pptx>=1.0.2",
    "pyyaml>=6.0.3",
]

[project.scripts]
ppt-creator = "ppt_creator:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
